+++ spec/javascripts/options/error_image_spec.js
@@ -9,22 +9,18 @@ describe('errorImage', function() {
     it ('appears on step title', function() {
       // given
       var
-        self   = $('form').stepy({ errorImage: true, validate: true }),
+        self = $('form').stepy({
+          errorImage: true,
+
+          validate: function(field) {
+            return self.validaty('validate', $(field)).data('valid');
+          }
+        }).validaty(),
         steps  = self.children('fieldset'),
         titles = $('#' + self.attr('id') + '-header').children('li');
 
-      self.validate({
-        errorPlacement: function(error, element) {
-          $('#stepy div.stepy-errors').append(error);
-        }, rules: {
-          'password':  'required'
-        }, messages: {
-          'password':  { required: 'Password field is requerid!' }
-        }
-      });
-
       // when
-      steps.eq(0).find('.stepy-next').click();
+      steps.eq(0).find('.stepy-next').trigger('click');
 
       // then
       expect(titles.eq(0)).toHaveClass('stepy-error');